requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai>=1.12.0
elevenlabs>=1.0.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=23.2.1
